[Refactor] Properties 패키지 생성 후 위치 옮김
