+++ src/main/java/com/catchsolmind/cheongyeonbe/domain/eraser/service/EraserService.java
@@ -20,7 +20,7 @@ import com.catchsolmind.cheongyeonbe.domain.user.entity.User;
 import com.catchsolmind.cheongyeonbe.domain.user.repository.UserRepository;
 import com.catchsolmind.cheongyeonbe.global.BusinessException;
 import com.catchsolmind.cheongyeonbe.global.ErrorCode;
-import com.catchsolmind.cheongyeonbe.global.config.S3Properties;
+import com.catchsolmind.cheongyeonbe.global.properties.S3Properties;
 import com.catchsolmind.cheongyeonbe.global.enums.SuggestionType;
 import com.catchsolmind.cheongyeonbe.global.enums.TaskStatus;
 import lombok.RequiredArgsConstructor;
